refactored sane_partial into utils/helpers.py
